requirements.txt
python-dotenv
# Polymarket BTC Dashboard - Dependencies
# Install with: pip install -r requirements.txt

requests>=2.28.0
websocket-client>=1.4.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
